fix StorageViewModel instantiation error

+++ BundleTransport/app/src/main/java/net/discdd/bundletransport/viewmodels/StorageViewModel.kt
@@ -99,9 +99,9 @@ class StorageViewModel(application: Application) : AndroidViewModel(application)
                 .apply()
     }
 
-    private fun retrievePreference(): Int {
+    private fun retrievePreference(): Long {
         return context.getSharedPreferences("SeekBarPrefs", Context.MODE_PRIVATE)
-                .getInt("seekBarPosition", 0)
+                .getLong("seekBarPosition", 0)
     }
 
     private fun getTotalBytes(): Long {
